[#21] 앨범에서 이미지 가져오기
